requirements.txt
aiomysql==0.3.2
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
idna==3.11
mysql==0.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
parameterized==0.9.0
pycodestyle==2.14.0
PyJWT==2.10.1
PyMySQL==1.1.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
